+++ app/dashboard/page.tsx
@@ -1,8 +1,5 @@
-'use client';
-
-import React from 'react';
-
-const RouteManagementPage: React.FC = () => {
+// app/dashboard/page.tsx
+export default function DashboardPage() {
   return (
     <div
       style={{
@@ -11,10 +8,9 @@ const RouteManagementPage: React.FC = () => {
         marginTop: '100px',
         color: '#000'
       }}
-    >
-      <h1>This is dedicated for Route Management</h1>
+      >
+      <h1>This is the Dashboard Page</h1>
+      {/* Just your page content */}
     </div>
   );
-};
-
-export default RouteManagementPage;
+}
