--- app/page.tsx
@@ -1,16 +1,0 @@
-// app/dashboard/page.tsx
-export default function DashboardPage() {
-  return (
-    <div
-      style={{
-        fontFamily: 'Arial, sans-serif',
-        textAlign: 'center',
-        marginTop: '100px',
-        color: '#000'
-      }}
-      >
-      <h1>This is the Dashboard Page</h1>
-      {/* Just your page content */}
-    </div>
-  );
-}
